+++ lib/SKGB/Intern/Controller/Export.pm
@@ -36,7 +36,7 @@ QUERY
 sub intern1 {
 	my ($self) = @_;
 	
-	if ( ! $self->has_access ) {
+	if ( ! $self->skgb->may ) {
 		return $self->render(template => 'key_manager/forbidden', status => 403);
 	}
 	
@@ -48,7 +48,7 @@ sub intern1 {
 sub listen {
 	my ($self) = @_;
 	
-	if ( ! $self->has_access ) {
+	if ( ! $self->skgb->may ) {
 		return $self->render(template => 'key_manager/forbidden', status => 403);
 	}